+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ## [unreleased]
 
+### Changed
+
+- `sandstone/PageViews` styling to match the latest GUI
+
 ### Fixed
 
 - `sandstone/Input` back button to be disabled when `disabled` prop is `true`