+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 - `sandstone/Scroller` `editable.blurItemFuncRef` prop to provide a function for blurring the focused item
 
+### Fixed
+
+- `sandstone/Input` back button to be disabled when `disabled` prop is `true`
+
 ## [2.9.0-beta.1] - 2024-06-17
 
 ### Added