+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Fixed
 
+- `sandstone/Popup` to restore focus when a popup is open by default and then closed
 - `sandstone/Input` back button to be disabled when `disabled` prop is `true`
 
 ## [2.9.0-beta.1] - 2024-06-17