+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 - `sandstone/Scroller` `editable.blurItemFuncRef` prop to provide a function for blurring the focused item
 
+### Changed
+
+- `sandstone/PageViews` styling to match the latest GUI
+
 ### Fixed
 
 - `sandstone/Input` back button to be disabled when `disabled` prop is `true`