+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `sandstone/Scroller` `editable.blurItemFuncRef` prop to provide a function for blurring the focused item
+
+### Changed
+
+- `sandstone/PageViews` styling to match the latest GUI
+
+### Fixed
+
+- `sandstone/Input` back button to be disabled when `disabled` prop is `true`
+
 ## [2.9.0-beta.1] - 2024-06-17
 
 ### Added