+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [2.7.17] - 2024-07-08
+
+### Added
+
+- `sandstone/Icon` supported icon list, adding new icon `create`
+- `sandstone/Icon` supported icon list, adding new wifi5G icons
+- `sandstone/Scroller` `editable.blurItemFuncRef` prop to provide a function for blurring the focused item
+
 ## [2.7.16] - 2024-05-13
 
 ### Added